function org
moved all tested parses into reusable functions that do chunks of parsing or file writing/ pandas DF adjustments. There are still things to perfect and a few formulas to finish. But this code works to get the data into the CSV in the format I am currently thinking will work well in SQL.